golangci-lint phase 4
clean up some final linter issues and add a make target for
golangci-lint. in addition, begin running the tests are part of the
gating tasks in cirrus ci.

we cannot fully shift over to the new linter until we fix the image on
the openshift side.  for short term, we will use both

+++ pkg/util/utils_linux.go
@@ -39,8 +39,8 @@ func FindDeviceNodes() (map[string]string, error) {
 		if !ok {
 			return errors.Errorf("Could not convert stat output for use")
 		}
-		major := uint64(sysstat.Rdev / 256)
-		minor := uint64(sysstat.Rdev % 256)
+		major := sysstat.Rdev / 256
+		minor := sysstat.Rdev % 256
 
 		nodes[fmt.Sprintf("%d:%d", major, minor)] = path
